requirements.txt
openpyxl==3.1.2
Pillow==10.4.0
st_pages==1.0.1
streamlit==1.31.1
streamlit_js_eval==0.1.5
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
